requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pandas
python-dotenv
pyyaml
requests
